pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "fast_prompt"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "taindp98" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
requires-python = "~=3.9"

# Add this section for runtime dependencies
dependencies = [
    "requests==2.32.3",
    "python-dotenv",
    "openai==1.50.2"
]

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff.lint.isort]
known_first_party = ["fast_prompt"]
force_sort_within_sections = true
